Rename hdf utilities file
