+++ src/toast/io/CMakeLists.txt
@@ -5,6 +5,6 @@ install(FILES
     __init__.py
     observation_hdf_save.py
     observation_hdf_load.py
-    observation_hdf_utils.py
+    hdf_utils.py
     DESTINATION ${PYTHON_SITE}/toast/io
 )
